Introduce common testing package for watch tests
