+++ pkg/virt-controller/watch/BUILD.bazel
@@ -144,6 +144,7 @@ go_test(
         "//pkg/virt-controller/watch/descheduler:go_default_library",
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
         "//pkg/virt-controller/watch/drain/evacuation:go_default_library",
+        "//pkg/virt-controller/watch/testing:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-controller/watch/util:go_default_library",
         "//staging/src/kubevirt.io/api/clone/v1alpha1:go_default_library",
