+++ pkg/virt-controller/watch/replicaset_test.go
@@ -31,6 +31,7 @@ import (
 	virtcontroller "kubevirt.io/kubevirt/pkg/controller"
 	"kubevirt.io/kubevirt/pkg/pointer"
 	"kubevirt.io/kubevirt/pkg/testutils"
+	watchtesting "kubevirt.io/kubevirt/pkg/virt-controller/watch/testing"
 )
 
 var _ = Describe("Replicaset", func() {
@@ -407,7 +408,7 @@ var _ = Describe("Replicaset", func() {
 			rs.Status.Replicas = 1
 			rs.Status.ReadyReplicas = 1
 			vmi.Status.Phase = v1.Running
-			markAsReady(vmi)
+			watchtesting.MarkAsReady(vmi)
 
 			addReplicaSet(rs)
 			addVMI(vmi)
@@ -450,7 +451,7 @@ var _ = Describe("Replicaset", func() {
 			// Move one VirtualMachineInstance to a final state
 			modifiedVMI := vmi.DeepCopy()
 			modifiedVMI.Status.Phase = v1.Running
-			markAsReady(modifiedVMI)
+			watchtesting.MarkAsReady(modifiedVMI)
 			modifiedVMI.ResourceVersion = "1"
 			vmiFeeder.Modify(modifiedVMI)
 
@@ -466,7 +467,7 @@ var _ = Describe("Replicaset", func() {
 			rs.Status.Replicas = 1
 			rs.Status.ReadyReplicas = 1
 			vmi.Status.Phase = v1.Running
-			markAsReady(vmi)
+			watchtesting.MarkAsReady(vmi)
 			addReplicaSet(rs)
 			addVMI(vmi)
 
@@ -478,7 +479,7 @@ var _ = Describe("Replicaset", func() {
 
 			// Move one VirtualMachineInstance to a final state
 			modifiedVMI := vmi.DeepCopy()
-			markAsNonReady(modifiedVMI)
+			watchtesting.MarkAsNonReady(modifiedVMI)
 			modifiedVMI.ResourceVersion = "1"
 			vmiFeeder.Modify(modifiedVMI)
 
@@ -519,7 +520,7 @@ var _ = Describe("Replicaset", func() {
 			rs.Status.Replicas = 1
 			rs.Status.ReadyReplicas = 1
 			vmi.Status.Phase = v1.Running
-			markAsReady(vmi)
+			watchtesting.MarkAsReady(vmi)
 			addReplicaSet(rs)
 			addVMI(vmi)
 
@@ -558,7 +559,7 @@ var _ = Describe("Replicaset", func() {
 			rs.Status.Replicas = 1
 			rs.Status.ReadyReplicas = 1
 			vmi.Status.Phase = v1.Running
-			markAsReady(vmi)
+			watchtesting.MarkAsReady(vmi)
 			addReplicaSet(rs)
 			addVMI(vmi)
